PolyMutable with element
