+++ checker/src/main/java/org/checkerframework/checker/pico/PICOQualifierHierarchy.java
@@ -1,40 +1,15 @@
-package org.checkerframework.checker.pico;
-
-import org.checkerframework.checker.pico.qual.PICOBottom;
-import org.checkerframework.checker.pico.qual.PICOLost;
-import org.checkerframework.framework.type.GenericAnnotatedTypeFactory;
-import org.checkerframework.framework.type.NoElementQualifierHierarchy;
-
-import java.lang.annotation.Annotation;
-import java.util.Collection;
-
-import javax.lang.model.element.AnnotationMirror;
-import javax.lang.model.util.Elements;
-
-/** A qualifier hierarchy for the PICO checker. */
-public class PICOQualifierHierarchy extends NoElementQualifierHierarchy {
-
-    /**
-     * Creates a PICOQualifierHierarchy from the given classes.
-     *
-     * @param qualifierClasses classes of annotations that are the qualifiers
-     * @param elements element utils
-     * @param atypeFactory the associated type factory
-     */
-    public PICOQualifierHierarchy(
-            Collection<Class<? extends Annotation>> qualifierClasses,
-            Elements elements,
-            GenericAnnotatedTypeFactory<?, ?, ?, ?> atypeFactory) {
-        super(qualifierClasses, elements, atypeFactory);
-    }
-
-    @Override
-    public boolean isSubtypeQualifiers(AnnotationMirror subAnno, AnnotationMirror superAnno) {
-        // Lost is not reflexive and the only subtype is Bottom
-        if (atypeFactory.areSameByClass(superAnno, PICOLost.class)
-                && !atypeFactory.areSameByClass(subAnno, PICOBottom.class)) {
-            return false;
-        }
-        return super.isSubtypeQualifiers(subAnno, superAnno);
-    }
-}
+// package org.checkerframework.checker.pico;
+//
+// import org.checkerframework.checker.pico.qual.PICOBottom;
+// import org.checkerframework.checker.pico.qual.PICOLost;
+// import org.checkerframework.framework.type.ElementForPolyQualifierHierarchy;
+// import org.checkerframework.framework.type.GenericAnnotatedTypeFactory;
+// import org.checkerframework.framework.type.NoElementQualifierHierarchy;
+// import org.checkerframework.framework.type.QualifierHierarchy;
+//
+// import java.lang.annotation.Annotation;
+// import java.util.Collection;
+//
+// import javax.lang.model.element.AnnotationMirror;
+// import javax.lang.model.util.Elements;
+//
